+++ bio_embeddings/mutagenesis_visualize/__init__.py
@@ -1,8 +1,0 @@
-"""
-WIP Do Not Use
-
-Visualize the results of the in-silico mutagenesis stage using an interactive heatmap
-"""
-from bio_embeddings.mutagenesis_visualize.pipeline import plot, run
-
-__all__ = ["plot", "run"]