+++ bio_embeddings/visualize/__init__.py
@@ -1,5 +1,6 @@
 """Display t-SNE or UMAP projections from the project stage"""
 
+from bio_embeddings.visualize.mutagenesis import plot_mutagenesis
 from bio_embeddings.visualize.plotly_plots import (
     render_3D_scatter_plotly,
     render_scatter_plotly,
@@ -7,6 +8,7 @@ from bio_embeddings.visualize.plotly_plots import (
 )
 
 __all__ = [
+    "plot_mutagenesis",
     "render_3D_scatter_plotly",
     "render_scatter_plotly",
     "save_plotly_figure_to_html",