+++ test-data/mutagenesis/plot_mutagenesis.py
@@ -3,13 +3,13 @@ from pathlib import Path
 import pandas
 import plotly
 
-from bio_embeddings.mutagenesis_visualize.pipeline import plot
+from bio_embeddings.visualize.mutagenesis import plot_mutagenesis
 
 
 def main():
     cwd = Path(__file__).resolve().parent
     probabilities = pandas.read_csv(cwd.joinpath("probabilities.csv"))
-    plotly.offline.plot(plot(probabilities), filename=str(cwd.joinpath("plot.html")))
+    plotly.offline.plot(plot_mutagenesis(probabilities), filename=str(cwd.joinpath("plot.html")))
 
 
 if __name__ == "__main__":